Few minor formatting changes
