+++ Python_Files/main.py
@@ -3,10 +3,12 @@ import problems.multiply_algorithm as multiply_algorithm
 import problems.division_algorithm as division_algorithm
 
 
+# Determines whether system is unix or windows based,
+# then sets appropriate clear for terminal
 def clearTerminalWindow():
-    if os.name == "nt":
+    if os.name == "nt":  # Windows based
         _ = os.system("cls")
-    else:
+    else:  # Unix based
         _ = os.system("clear")
 
 
@@ -26,8 +28,8 @@ def mainMenuLoop():
     while not userInput == "q":
 
         print(title)
-        print("<1> Multiply")
-        print("<2> Divide")
+        print("<1> Multiply 4-bit")
+        print("<2> Divide 4-bit")
         print("\n<q> Quit")
         userInput = input("enter choice:")
 
@@ -38,46 +40,6 @@ def mainMenuLoop():
         else:
             clearTerminalWindow()
             continue
-            
 
-# divAlg(4, 4)
-# multAlg(4, 4)
-                
-
-
-
-
-
-
-
-
-
-
-class MipsInstruct:
-    functType = ""
-    opcode = ""
-    mnemonic = ""
-    # this should always follow this order rd, rs, rt
-    registers = ["", "", ""]
-    shamt = ""
-    funct = ""
-    imm = ""
-
-
-# basic machine to mips game
-def machToMips():
-    # define dictionaries for different parts of mip instructions to be translated
-    # dictionary instrType = list out all possible for core instr set
-    # i.e.
-    # dictionary registers = pair all registers to corresponding number
-    # dictionary mnemonics = this should be fed into from instr type dict
-
-    # define a function that will be able to choose a random value in a dictionary
-    # choose random value for
-    return
-
-
-# divAlg(4, 4)
-# multAlg(4, 4)
 
 mainMenuLoop()
